backend/requirements.txt
fastapi
uvicorn[standard]
celery[redis]
redis
sqlalchemy
pydantic
python-socketio
